Slight advantage from summing doubles.

+++ orogram/xentropy.pxd
@@ -5,4 +5,4 @@
 # cython: wraparound=False
 # cython: nonecheck=False
 
-cpdef float section_crossentropy(float p0, float p1, float q0, float q1, double log_q0, double log_q1) nogil
+cpdef double section_crossentropy(float p0, float p1, float q0, float q1, double log_q0, double log_q1) nogil
